cd device_fallback
This commit imports debian patch "42_cd_device_fallback.patch".

If the user doesn't specify a device and the default (/dev/cdrom) doesn't
exist, it would be nice if jack could fall back to the device(s) listed in
/proc/sys/dev/cdrom/info (at least on Linux).  According to the
linux-kernel mailing list, this is a stable interface.

Debian #353551
Depends on 21_fallback_ripper.patch