enable term
This commit imports debian patch "40_enable_term.patch".

When a ripper is killed with ctrl-c, it is properly marked as "DAE failed
with status 15, wav removed".  However, afterwards pressing any key
does precisely nothing, e.g. cursor keys or ctrl-l.  ctrl-c does work,
but messes up the display.

The reason for that is that jack_term.disable() is called in the signal
handler, but nobody calls jack_term.enable()... adding this works.

In addition, let's ignore ctrl-c in the signal handler function because
otherwise people could press ctrl-c again within the signal handler (e.g.
while jack is waiting for the user to press enter when --wait is on), and
then we get the display problem again because jack_term.enable() is not
called.

Debian #352755

P.S. I'm not quite sure whether this fix (and 56_ctrl_c_termset.patch,
which is related) are correct.  Basically, according to the code, pressing
Ctrl-C should clean up and then exit the program.  However, for some reason
jack never exits; all that happens is that the helpers are killed off.  The
sig variable seems to indicate whether we're really exiting or not...
unfortunately, I have no idea why jack doesn't exit all the time or what
the correct behaviour should be.  See http://bugs.debian.org/196429 for
some discussion about this.

Depends on 21_fallback_ripper.patch