man minus
This commit imports debian patch "32_man_minus.patch".

Escape minus signs in the man page properly.

'-' must be escaped ('\-') to be interpreted as minus. If you really
intend a hyphen, write it as '\(hy' to emphasise that fact. See
groff(7) and especially groff_char(7) for details, and also the thread
starting with
http://lists.debian.org/debian-devel/2003/debian-devel-200303/msg01481