nicify wait
This commit imports debian patch "41_nicify_wait.patch".

When --wait is on jack always display "press ENTER to exit" when leaving
the signal handler.  This message is not optimal since it's also displayed
when you e.g. kill a ripper or encoder.  Check whether sig is set and adapt
the message accordingly.

Depends on 40_enable_term.patch