feat: remove setuptools_scm from REQUIRES (#19)
also, reverse the order of setuptools_scm and importlib.metadata
when reading the version.

setuptools_scm is now tried first, as it tries to get the version from
git directly, which is more up to date while developing.

if setuptools_scm is not available or if jack is started from outside
of git, fall back on importlib.metadata.